remote_address: refactor Path to its own source file
