+++ src/lib.rs
@@ -8,6 +8,7 @@ mod operation_dispatcher;
 mod policy;
 mod policy_index;
 mod property;
+mod property_path;
 mod service;
 
 #[cfg(test)]
