+++ src/property_path.rs
@@ -1,0 +1,89 @@
+use std::fmt::{Debug, Display, Formatter};
+
+#[derive(Debug, Clone)]
+pub struct Path {
+    tokens: Vec<String>,
+}
+
+impl Display for Path {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            self.tokens
+                .iter()
+                .map(|t| t.replace('.', "\\."))
+                .collect::<Vec<String>>()
+                .join(".")
+        )
+    }
+}
+
+impl From<&str> for Path {
+    fn from(value: &str) -> Self {
+        let mut token = String::new();
+        let mut tokens: Vec<String> = Vec::new();
+        let mut chars = value.chars();
+        while let Some(ch) = chars.next() {
+            match ch {
+                '.' => {
+                    tokens.push(token);
+                    token = String::new();
+                }
+                '\\' => {
+                    if let Some(next) = chars.next() {
+                        token.push(next);
+                    }
+                }
+                _ => token.push(ch),
+            }
+        }
+        tokens.push(token);
+
+        Self { tokens }
+    }
+}
+
+impl Path {
+    pub fn tokens(&self) -> Vec<&str> {
+        self.tokens.iter().map(String::as_str).collect()
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+
+    #[test]
+    fn path_tokenizes_with_escaping_basic() {
+        let path: Path = r"one\.two..three\\\\.four\\\.\five.".into();
+        assert_eq!(
+            path.tokens(),
+            vec!["one.two", "", r"three\\", r"four\.five", ""]
+        );
+    }
+
+    #[test]
+    fn path_tokenizes_with_escaping_ends_with_separator() {
+        let path: Path = r"one.".into();
+        assert_eq!(path.tokens(), vec!["one", ""]);
+    }
+
+    #[test]
+    fn path_tokenizes_with_escaping_ends_with_escape() {
+        let path: Path = r"one\".into();
+        assert_eq!(path.tokens(), vec!["one"]);
+    }
+
+    #[test]
+    fn path_tokenizes_with_escaping_starts_with_separator() {
+        let path: Path = r".one".into();
+        assert_eq!(path.tokens(), vec!["", "one"]);
+    }
+
+    #[test]
+    fn path_tokenizes_with_escaping_starts_with_escape() {
+        let path: Path = r"\one".into();
+        assert_eq!(path.tokens(), vec!["one"]);
+    }
+}
